Adopt ruff: config, a check.sh gate, a CI job, and the 58 findings it reported
The RUF012 request had come up three times and been declined three times on
the grounds that nothing in the repo runs ruff, which was true and is the part
worth fixing rather than restating.

## The gate

`ruff.toml` at the root pins `target-version = "py39"` — the floor
`check.sh` actually runs on, since it invokes whatever `python3` is on PATH —
and an explicit `select`, because ruff's default set has widened across
releases and a rule set that moves when the tool is upgraded reports different
things on two machines.

`tools/check.sh ruff` runs it and is part of `fast`. When ruff is absent it
*fails* with install instructions rather than skipping, for the reason this
branch keeps arriving at: a check that quietly does nothing still shows a
green tick. A `python` job in CI installs ruff, runs the lint, and also runs
`test-bench` — which had never run in CI at all, so the harness that produces
every device number this project has was gated only by whoever remembered to
run it locally.

Four rule families are ignored, each with its reason in the config: RUF001-003
(the prose uses em dashes and typographic quotes deliberately) and RUF005
(`a + [b]` reads as "this run plus this event"; the unpacking form obscures
it). Nothing else is suppressed wholesale.

## The findings

58, all fixed rather than ignored, except one:

- RUF012 (22) — the original request. Every class-level fixture in
  test_bench.py is annotated, not the two that happened to be asked about.
  `from __future__ import annotations` comes with them, so the annotations
  cost nothing at runtime and `X | None` works on 3.9.
- SIM115 (9) — leaked file handles. convert.py's eight one-shot reads and
  writes become `Path.read_text`/`write_text`, which also stops them writing
  in the platform's default encoding while reading UTF-8.
- ISC004 (4) and SIM117 (5) — implicit concatenation inside list literals,
  the missing-comma footgun a reviewer caught by hand earlier on this branch,
  and nested `with`. The flattened ones use `with a, b:` rather than the
  parenthesised group: 3.9's parser accepts that form by accident, and this
  repo's whole argument is against depending on interpreter accidents.
- RUF013 (3) — `seen: dict = None` and friends, genuinely wrong annotations.
- I001 (10), UP035, UP031, F401, RUF046 (3), RUF059, E741 (2), E402, B904 —
  import order, `collections.abc` over `typing`, percent formatting,
  an unused import, `int(round(...))`, an unused unpacked field, `l` as a
  name, a test import stranded mid-file, and a `raise` that dropped its cause.

The exception is `serial_capture.py`'s log handle, which is deliberately open
for the life of the process and appended from two places, so there is no scope
a context manager could close it at. It carries a narrow `# noqa: SIM115` with
that reason next to it, which is the rule AGENTS.md already sets for Rust
lints.

Not adopted: `ruff format`. It would reflow every Python file in the repo,
which is a separate decision from turning the linter on.

Verification: ruff clean across the repo on 0.16.1; check.sh ruff exits 1 with
install instructions when ruff is absent and 0 when it is present; 239 bench
tests on 3.9.6 and 3.14.6; stack_frames' own 16 tests; check.sh fmt,
clippy-host, test-host, test-host-x3; every touched script byte-compiles; and
ci.yml parses with the new job carrying the same pinned checkout SHA as its
neighbours. No Rust behaviour changed, so the firmware and emulator suites
were not re-run.

+++ tools/fontgen_common.py
@@ -1,9 +1,8 @@
-from pathlib import Path
 import os
 import struct
+from pathlib import Path
 
 from PIL import Image, ImageDraw
-
 
 ADVANCE_SCALE = 16
 DEFAULT_THRESHOLD = 128
@@ -26,7 +25,7 @@ THRESHOLD = text_render_threshold()
 
 
 def advance_fp(font, text: str) -> int:
-    return max(int(round(font.getlength(text) * ADVANCE_SCALE)), ADVANCE_SCALE)
+    return max(round(font.getlength(text) * ADVANCE_SCALE), ADVANCE_SCALE)
 
 
 def rasterize_glyph(font, code: int):
@@ -337,7 +336,7 @@ def kerning_entries(font_path: Path, cps, px: int):
 
     entries = []
     for (left_glyph, right_glyph), value in adjustments.items():
-        adjust_fp = int(round(value * px * ADVANCE_SCALE / units_per_em))
+        adjust_fp = round(value * px * ADVANCE_SCALE / units_per_em)
         if abs(adjust_fp) < MIN_KERNING_ADJUST_FP:
             continue
         for left in cps_by_glyph.get(left_glyph, []):
